+++ pdata/plog/json.go
@@ -16,11 +16,14 @@ package plog // import "go.opentelemetry.io/collector/pdata/plog"
 
 import (
 	"bytes"
+	"fmt"
 
 	"github.com/gogo/protobuf/jsonpb"
+	jsoniter "github.com/json-iterator/go"
 
 	"go.opentelemetry.io/collector/pdata/internal"
 	otlplogs "go.opentelemetry.io/collector/pdata/internal/data/protogen/logs/v1"
+	"go.opentelemetry.io/collector/pdata/internal/json"
 	"go.opentelemetry.io/collector/pdata/internal/otlp"
 )
 
@@ -45,7 +48,6 @@ func (e *jsonMarshaler) MarshalLogs(ld Logs) ([]byte, error) {
 }
 
 type jsonUnmarshaler struct {
-	delegate jsonpb.Unmarshaler
 }
 
 // NewJSONUnmarshaler returns a model.Unmarshaler. Unmarshals from OTLP json bytes.
@@ -54,14 +56,122 @@ func NewJSONUnmarshaler() Unmarshaler {
 }
 
 func newJSONUnmarshaler() *jsonUnmarshaler {
-	return &jsonUnmarshaler{delegate: jsonpb.Unmarshaler{}}
+	return &jsonUnmarshaler{}
 }
 
 func (d *jsonUnmarshaler) UnmarshalLogs(buf []byte) (Logs, error) {
-	ld := otlplogs.LogsData{}
-	if err := d.delegate.Unmarshal(bytes.NewReader(buf), &ld); err != nil {
-		return Logs{}, err
+	iter := jsoniter.ConfigFastest.BorrowIterator(buf)
+	defer jsoniter.ConfigFastest.ReturnIterator(iter)
+	ld := d.readLogsData(iter)
+	if iter.Error != nil {
+		return Logs{}, iter.Error
 	}
 	otlp.MigrateLogs(ld.ResourceLogs)
 	return Logs(internal.LogsFromProto(ld)), nil
+}
+
+func (d *jsonUnmarshaler) readLogsData(iter *jsoniter.Iterator) otlplogs.LogsData {
+	ld := otlplogs.LogsData{}
+	iter.ReadObjectCB(func(iter *jsoniter.Iterator, f string) bool {
+		switch f {
+		case "resource_logs", "resourceLogs":
+			iter.ReadArrayCB(func(iterator *jsoniter.Iterator) bool {
+				ld.ResourceLogs = append(ld.ResourceLogs, d.readResourceLogs(iter))
+				return true
+			})
+		default:
+			iter.Skip()
+		}
+		return true
+	})
+	return ld
+}
+
+func (d *jsonUnmarshaler) readResourceLogs(iter *jsoniter.Iterator) *otlplogs.ResourceLogs {
+	rs := &otlplogs.ResourceLogs{}
+	iter.ReadObjectCB(func(iter *jsoniter.Iterator, f string) bool {
+		switch f {
+		case "resource":
+			json.ReadResource(iter, &rs.Resource)
+		case "scope_logs", "scopeLogs":
+			iter.ReadArrayCB(func(iter *jsoniter.Iterator) bool {
+				rs.ScopeLogs = append(rs.ScopeLogs,
+					d.readScopeLogs(iter))
+				return true
+			})
+		case "schemaUrl", "schema_url":
+			rs.SchemaUrl = iter.ReadString()
+		default:
+			iter.Skip()
+		}
+		return true
+	})
+	return rs
+}
+
+func (d *jsonUnmarshaler) readScopeLogs(iter *jsoniter.Iterator) *otlplogs.ScopeLogs {
+	ils := &otlplogs.ScopeLogs{}
+	iter.ReadObjectCB(func(iter *jsoniter.Iterator, f string) bool {
+		switch f {
+		case "scope":
+			json.ReadScope(iter, &ils.Scope)
+		case "log_records", "logRecords":
+			iter.ReadArrayCB(func(iter *jsoniter.Iterator) bool {
+				ils.LogRecords = append(ils.LogRecords, d.readLog(iter))
+				return true
+			})
+		case "schemaUrl", "schema_url":
+			ils.SchemaUrl = iter.ReadString()
+		default:
+			iter.Skip()
+		}
+		return true
+	})
+	return ils
+}
+
+func (d *jsonUnmarshaler) readLog(iter *jsoniter.Iterator) *otlplogs.LogRecord {
+	lr := &otlplogs.LogRecord{}
+	iter.ReadObjectCB(func(iter *jsoniter.Iterator, f string) bool {
+		switch f {
+		case "timeUnixNano", "time_unix_nano":
+			lr.TimeUnixNano = json.ReadUint64(iter)
+		case "observed_time_unix_nano", "observedTimeUnixNano":
+			lr.ObservedTimeUnixNano = json.ReadUint64(iter)
+		case "severity_number", "severityNumber":
+			lr.SeverityNumber = d.readSeverityNumber(iter)
+		case "severity_text", "severityText":
+			lr.SeverityText = iter.ReadString()
+		case "body":
+			iter.ReadObjectCB(func(iter *jsoniter.Iterator, f string) bool {
+				lr.Body = json.ReadAnyValue(iter, f)
+				return true
+			})
+		case "attributes":
+			iter.ReadArrayCB(func(iter *jsoniter.Iterator) bool {
+				lr.Attributes = append(lr.Attributes, json.ReadAttribute(iter))
+				return true
+			})
+		case "droppedAttributesCount", "dropped_attributes_count":
+			lr.DroppedAttributesCount = json.ReadUint32(iter)
+		case "flags":
+			lr.Flags = json.ReadUint32(iter)
+		case "traceId", "trace_id":
+			if err := lr.TraceId.UnmarshalJSON([]byte(iter.ReadString())); err != nil {
+				iter.ReportError("readLog.traceId", fmt.Sprintf("parse trace_id:%v", err))
+			}
+		case "spanId", "span_id":
+			if err := lr.SpanId.UnmarshalJSON([]byte(iter.ReadString())); err != nil {
+				iter.ReportError("readLog.spanId", fmt.Sprintf("parse span_id:%v", err))
+			}
+		default:
+			iter.Skip()
+		}
+		return true
+	})
+	return lr
+}
+
+func (d *jsonUnmarshaler) readSeverityNumber(iter *jsoniter.Iterator) otlplogs.SeverityNumber {
+	return otlplogs.SeverityNumber(json.ReadEnumValue(iter, otlplogs.SeverityNumber_value))
 }